DDPB-2635 delete object versions and markers

+++ src/AppBundle/Service/File/Storage/S3Storage.php
@@ -105,12 +105,6 @@ class S3Storage implements StorageInterface
      */
     public function removeFromS3($key)
     {
-        // deleteObject inserts a delete marker against the current version
-        $this->s3Client->deleteObject([
-            'Bucket' => $this->bucketName,
-            'Key' => $key
-        ]);
-
         /*
          * ListObjectVersions is permitted by ListBucketVersions in IAM.
          */
@@ -121,13 +115,25 @@ class S3Storage implements StorageInterface
 
         if ($objectVersions instanceof \Aws\Result) {
             foreach ($objectVersions['Versions'] as $versionData) {
-                $this->s3Client->deleteObject([
-                    'Bucket' => $this->bucketName,
-                    'Key' => $versionData['Key'],
-                    'VersionId' => $versionData['VersionId'],
-                ]);
+                if (!empty($versionData["VersionId"])) {
+                    $this->s3Client->deleteObject([
+                        'Bucket' => $this->bucketName,
+                        'Key' => $versionData['Key'],
+                        'VersionId' => $versionData['VersionId'],
+                    ]);
+                }
             }
 
+            // remove any deleteMarkers permanently
+            foreach ($objectVersions['DeleteMarkers'] as $dmData) {
+                if (!empty($dmData["VersionId"])) {
+                    $this->s3Client->deleteObject([
+                        'Bucket' => $this->bucketName,
+                        'Key' => $dmData['Key'],
+                        'VersionId' => $dmData['VersionId'],
+                    ]);
+                }
+            }
             return true;
         }
